Added file with definition mime types
Optimized use of threads for processing requests
Minor optimizations in the parser configuration
