+++ main.conf
@@ -2,4 +2,6 @@ threads_max_count 512;
 
 default_temp_dir /tmp/;
 
+mimes mimes.conf;
+
 include apps.conf;
