+++ mimes.conf
@@ -1,0 +1,9 @@
+text/html			html
+text/javascript		js
+text/css			css
+
+image/png			png
+image/jpeg			jpg jpeg
+
+video/webm			webm
+video/mp4			mp4
